# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/inventory/movements": {
      "get": {
        "operationId": "InventoryController_movements_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["inventory"]
      }
    },
...
    "/api/v1/reports/sales": {
      "get": {
        "operationId": "ReportsController_sales_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["reports"]
      }
    },
    "/api/v1/reports/inventory/low-stock": {
      "get": {
        "operationId": "ReportsController_lowStock_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["reports"]
      }
    },
    "/api/v1/reports/inventory/movements": {
      "get": {
        "operationId": "ReportsController_movements_v1",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "itmarket_staff_access": [] }],
        "tags": ["reports"]
      }
    },
...
  },
  "info": {
    "title": "ITMarket API",
    "description": "ITMarket service API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "itmarket_customer_session",
        "description": "Opaque, HTTP-only customer session cookie"
      }
    },
...
  }
}